// src/app/notes/models/notes.model.ts
export interface Note {
  _id: string;
  date: Date;
  company: string;
  project: string;
  technology: string;
  title: string;
  subjectArea: string;
  detail: string;
  createdBy: string;
  course: string;
  courseSection: string;
  gitRepo: string;
  gitCommit: string;
  htmlContent: string;
  content: string;
  time: Date;
  user: any;
}